% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescoveragedeploy-package.R
\docType{_PACKAGE}
\name{bayescoveragedeploy-package}
\alias{bayescoveragedeploy-package}
\alias{bayescoveragedeploy}
\title{The 'bayescoveragedeploy' package.}
\description{
Provides precompiled Stan models for fitting local country-level
Bayesian hierarchical transition models for health coverage indicators
(ANC4, institutional delivery, vaccination). This package uses the rstan
backend with precompiled models for users without C++ compilers.
All core functionality is provided by the bayescoveragemodel package.
}
\details{
This package contains precompiled Stan models for the following variants:
\itemize{
\item fpem - Basic transition model
\item fpem_routine - Model with routine data
\item fpem_aggregates - Model with subnational aggregates
\item fpem_routine_aggregates - Model with routine data and aggregates
}

The main user-facing function is \code{\link{fit_local_model}}.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.7. https://mc-stan.org
}
