% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_local_model.R
\name{determine_model_variant}
\alias{determine_model_variant}
\title{Determine which Stan model variant to use}
\usage{
determine_model_variant(routine_df = NULL, add_aggregates = FALSE)
}
\arguments{
\item{routine_df}{Optional routine data}

\item{add_aggregates}{Whether aggregates are used}
}
\value{
Model name string
}
\description{
Determine which Stan model variant to use
}
\keyword{internal}
