% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_local_model.R
\name{get_global_fit}
\alias{get_global_fit}
\title{Get global fit for indicator}
\usage{
get_global_fit(indicator)
}
\arguments{
\item{indicator}{Indicator name}
}
\value{
Global fit object
}
\description{
Get global fit for indicator
}
\keyword{internal}
