% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_local_model.R
\name{get_precompiled_model}
\alias{get_precompiled_model}
\title{Get precompiled Stan model}
\usage{
get_precompiled_model(model_name)
}
\arguments{
\item{model_name}{Name of the Stan model}
}
\value{
Precompiled Stan model object
}
\description{
Get precompiled Stan model
}
\keyword{internal}
