% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescoveragedeploy-package.R
\docType{package}
\name{bayescoveragedeploy-package}
\alias{bayescoveragedeploy-package}
\alias{bayescoveragedeploy}
\title{The 'bayescoveragedeploy' package.}
\description{
Provides precompiled Stan models for fitting local country-level
Bayesian hierarchical transition models for health coverage indicators
(ANC4, institutional delivery, vaccination). This package uses the rstan
backend with precompiled models for users without C++ compilers.
All core functionality is provided by the bayescoveragemodel package.
}
\details{
This package contains precompiled Stan models for the following variants:
\itemize{
\item fpem - Basic transition model
\item fpem_routine - Model with routine data
\item fpem_aggregates - Model with subnational aggregates
\item fpem_routine_aggregates - Model with routine data and aggregates
}

The main user-facing function is \code{\link{fit_local_model}}.
}
\section{Examples}{

Comprehensive examples are available in \code{inst/examples/}:
\itemize{
\item \code{basic_usage.R} - Complete tutorial with all features
\item \code{deployment_example.R} - Deployment workflow examples
\item \code{quick_reference.R} - Copy-paste snippets for common tasks
}

Run examples with:
\preformatted{
source(system.file("examples/basic_usage.R", package = "bayescoveragedeploy"))
}
}

\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.7. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlkemaLab/bayescoveragedeploy}
  \item Report bugs at \url{https://github.com/AlkemaLab/bayescoveragedeploy/issues}
}

}
\author{
\strong{Maintainer}: Leontine Alkema \email{lalkema@umass.edu}

Authors:
\itemize{
  \item Shauna Mooney
}

Other contributors:
\itemize{
  \item Sophia Kagoye [contributor]
  \item Leonardo Ferreira [contributor]
  \item Roland Mady [contributor]
  \item Countdown BayesCoverage Team [contributor]
}

}
